Homework over the holiday break
I pretty much got through most of it, although I am having challenges
with the header/nav bar background color.

I would also like to know how to make the width of the footer and
header extend all the way from the left to right instead of being boxed
into the width of the content.